+++ src-tauri/gen/schemas/desktop-schema.json
@@ -1,0 +1,1 @@
+{schema_str}
