+++ src-tauri/gen/schemas/linux-schema.json
@@ -1,0 +1,1 @@
+{schema_str}
